+++ internal/provider/types/publishers_private_apps_response_reachability.go
@@ -5,7 +5,7 @@ package types
 import "github.com/hashicorp/terraform-plugin-framework/types"
 
 type PublishersPrivateAppsResponseReachability struct {
-	ErrorCode   types.Int64  `tfsdk:"error_code"`
+	ErrorCode   types.Int32  `tfsdk:"error_code"`
 	ErrorString types.String `tfsdk:"error_string"`
 	Reachable   types.Bool   `tfsdk:"reachable"`
 }
